Removes Inhumen/Psydonite/Zizoid Novitiate (#2968)
## About The Pull Request

This PR is a very simple change: It prevents Novitiates from choosing
any patron other than the Divine Pantheon ones.

## Why It's Good For The Game

As it stands right now, Novitiate is a role that's a bit of an awkward
leftover from old Roguetown's young roles, pretty much obsoleted by
Acolyte. Still, it has a place as a learning role and something of a
church-assistant, although the church is never so crowded that you
wouldn't just be able to play Acolyte instead.

Still, the way it exists currently has a single, crucial flaw: there is
no patron restriction on Novitiate. Whether unintentional or intended, I
believe this to be a flaw in the role. It leads to the role being used
not as a learning or helper role, but as a saboteur role. People only
play as novitiate with the intention of being an inhumen or
Zizo-worshipper that can sabotage the church from the inside. It's a
completely baffling thing to have exist, and it's the only church role
that isn't restricted in terms of patron.

This simple change would make it so that these saboteurs actually have
to work to get their way into a trusted position from which they can
sabotage the church-- rather than being able to do so from roundstart.
It makes sense that the church would be a bit strict on who exactly gets
to wear the cloth, especially considering they all command miracles.

It also keeps a bit more consistency in the lore. Right now, Psydonites
can be novitaites. Even though lorewise, Psydon-worshippers are not
meant to have any sort of divine power, due to the way the role is
coded, Psydonite novitiates still get a functional miracle. With this
change, Psydonite novitiates will no longer be possible.

## Proof of Testing (Required)

https://www.youtube.com/watch?v=ICol28zaNv4