AP Greatweapon Strap - - - Additions:  - Includes a 'greatweapon strap', pulled directly from AP. - - - Roles Modified:  - Knight Sentinel and Champion provided the GWS.  - Retinue Halberdiers provided the GWS.  - Grenzelhoft Mercenaries(Excluding Jaeger) provided the GWS.  - Bandit Knave and Rogue Mage provided the GWS. - - - Crafting:  - You can craft the GWS, classified as a skill 3 sewing recipe(Apprentice). It requires 2 leather + 1 rope. - - -
What does this do? It allows weapons of a bulky size, such as halberds, estocs and spears to be strapped directly to your back. To draw and stow the weapons, it's a five second timer. This means it's almost entirely inadvisable to be doing this in combat. Stats, for what little there are with such an item, are also derived entirely from AP and unchanged. (The time required, more-or-less.)