Reverts drunkeness code back to previous state (#2978)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Complaints from players this might be a bit high. I'll try to figure out
how alcohol works on a deeper level before re-implementing this.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
3 beers = death right now.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/27a73f25-ed29-47ae-922d-546ca461ff85)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->