Elderbaum is angry! Makes dryad when hit. (#2902)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
When the elderbaum is hit every 60 seconds from its last spawn it will
spawn a dryad to persuade people to stop.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The elderbaum **will** be respected. Too many people decide to kill it
as a joke and face no consequence for it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

https://github.com/user-attachments/assets/72dfa042-655f-426d-982c-88dffca3b64a

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->