Bandit Additions - Kidnapping & Branding FIX ( Original is #2630) (#2881)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Promised content that bandits were supposed to get for their sins and
nerfs
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2630 but fixed

## Why It's Good For The Game

IS HOARDMASTER NERFED AS FUCK??
No. Wait 45 minutes since round start and 1 hour and 20 minutes from
round start. It will unlock FIRST and SECOND supply packs for you.


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->