Memory & Nerve Master job anonymisation improvements (#2905)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR aims to prevent metagaming by anonymising jobs of "foreigner"
roles in the nerve master and memory interfaces, alongside preventing
werewolves from appearing under their werewolf names in the nerve master
bank interface.

In cases where a character's examine text would show them as a wanderer
or foreigner (with the exception of mercenaries), these characters' jobs
will now display as "Foreigner" in the memory and nervemaster
interfaces, rather than revealing information that should be hidden (for
example, revealing somebody to be an Apostate). In the case of a
transformed werewolf, the nerve master will now show the werewolf's
regular name, if they have an account, rather than the werewolf name
assigned when they transform.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
These changes restrict two avenues of metagaming that I've become aware
that some players exploit - namely, that one can identify a refugee or
wanderer's role by examining them and checking their memories, or by
checking their account in the nervemaster, should they create one at a
nervelock. I'm not sure of the precise extent to which these issues were
exploited, but a friend has told me of one occasion on which they, as
Duke, were told by the Steward to declare several people Outlaws without
much explanation as to why, only later learning that the Steward had
been exploiting this oversight.

With these oversights fixed, it will no longer be possible to gain
knowledge of player roles or werewolf identities in this way, reducing
both the temptation and ability to metagame.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Nerve master interface showing my character's job as "FOREIGNER" after
joining as an Apostate Duallist:

![image](https://github.com/user-attachments/assets/c2cb7632-a983-4173-8758-b685db1ecd4d)

The same, with compact mode enabled:

![image](https://github.com/user-attachments/assets/33a4805c-ffcc-4800-9f2e-9e10b0d45d1b)

Having switched to another character and examined the original
character, my memories, listing the original character's job as
"Foreigner":

![image](https://github.com/user-attachments/assets/87af5e5e-d051-4250-909f-fd09af43961e)

Image showing my character's regular name, even after being transformed
into a werewolf:

![image](https://github.com/user-attachments/assets/3bcd69c1-0759-49f2-b44e-856d2864e7ff)

And confirmation that the "unknown" fallback still functions correctly,
when no job is available:

![image](https://github.com/user-attachments/assets/e18bdaaa-6531-4f6c-994f-1acc8718f06b)

As does the fallback in the nerve master, for those with an account but
no job:

![image](https://github.com/user-attachments/assets/36acd2eb-02ed-40cb-bbe5-b0d2dcef51e3)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->