Adds slipping procs, uses athletic and perception as checks (#2920)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
- Uses both the athletics as a main check for slipping and a backup
check for perception in case it fails as a fallback attempt to save you
from falling.
- Adds this new proc to soap as a proof of concept.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Allows us to expand slipping with more precision and get more use out of
some skills.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/397426f4-78bc-431f-b7ef-021c00185369


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->