allows pickpocketing from sleeping and closed eye humans (#2941)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
adds two or statements to pickpocket checks to see if the other person
has their eyes closed or is asleep. I don't think this caused any issues
with seeing a message with the eyes closed because if it's a visible
message neither of these situations should tip you off as far as I know.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes sense. They check if they can see you. They would not be able to
see you in either of these situations.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/4ede9e6b-8f9f-431d-8059-1a072c2302e5


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->