Divine Pantheon Novitiate Fix (#2985)
## About The Pull Request

This PR fixes the mistake I made in my previous PR (
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2968 ), pointed out by
Tastyfish which cuts out Xylix, Abyssor, and Necra from possible
Novitiate patron choices. Whoopsie!

## Why It's Good For The Game

Not really sure if anyone actually wants to play with those patrons in
this role, but it SHOULD ostensibly be allowed! No reason to leave
behind a mistake if it's an easy fix!

## Proof of Testing (Required)

N/A, same as before.