Tabaxi now have purr efect when saying anything with R.  (#2894)
## About The Pull Request

Tabaxi now have their own tongue, which affects their speach slightly.
Somewhat simillar to the old drakians, and perhaps better example would
be lizard people from TG.

If its a problem, I can also just do it via selectable accent. But that
list is only growing bigger and bigger, so I decided to stick with this.

## Why It's Good For The Game

Makes the way that tabaxi is speaking a bit more unique. Its similar to
what cat tongue had in the past on Skyrat, its just multiplying R or r
within the sentance, for slight purring effects. Thats the easiest way
to achive it, instead of replacing whole words.

This might not be liked by many people, but Im sure it adds a bit more
flavour overall to the race, and making them slightly more unique
compared to other races.


## Proof of Testing (Required)

Will add it soon, it works. But might refine the regex values a bit, not
sure of it yet.