Changes "jumpscare (1).ogg" sound to F&H battle start sfx (#2989)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The current jumpscare (1) sounds rough and stands out as the worst one.
Much more painful to listen to than the others. I changed it to be the
fear and hunger battle start sound instead. These are also only ever
used in grass ambushes so it's kind of like an RPG encounter.

sounds:


https://github.com/user-attachments/assets/c74f490c-5ccb-456d-bf04-106dcd63d0e3


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It's less painful on the ears perhaps.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
sound file replacement...
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->