Correctly equips zig butts to mouth (#2998)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Smoke butts would directly assign themselves to your mouth slot. While
this did technically work it would leave the slot graphically broken
until you clicked on it or spat out the butt. This fixes that to make it
equip in a way the inventory system likes.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
visual bug.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/06338fce-1c6e-4e2b-a56e-3b394283698c


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->