Adds find function to popup windows (#3006)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
allows you to ctrl+F in pop up windows to search for text. Was trying to
make it just work for known people but I guess it working on any pop up
is fine
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
trying to find some things like people you know can be hard if you got a
big list and can't parse it. This helps speed that process up a bit.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
proof of search in windows:

![image](https://github.com/user-attachments/assets/214cd95f-71ef-42c5-bc80-14a955b2a01d)


![image](https://github.com/user-attachments/assets/09e8814b-ecb2-47f7-8a04-fbb18d9e80f3)


![image](https://github.com/user-attachments/assets/7bed4565-3a3e-4be6-ac23-055027237f3a)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->