Slaver's License for Zybantine Party (#2992)
## About The Pull Request

Adds a slaver's license to the Zybantine wave.
Adds a copy of the tome of justice to the Zybantine wave.

## Why It's Good For The Game

Being an active slaver requires a license by the queen to be legitimate.
Zybantines were always supposed to have one, but we just never got
around to it until now.
Zybantines must still abide by the tome of justice to be within the
limits outlined in the tome.


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/86d3d288-d088-4915-a386-46dd5a4e20b3)