Refactors Gem information, creates arcane power global list, cleans remarks list. (#2952)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Refactor to organize gems a bit. They were in different files and had
variables assigned in them as well. Also turns arcane power into a list
in case it would ever need to be expanded. Changes the logic for reading
the list when using it in the spellbook code.

The remarks list chose to fight me so it is no longer a large single
line but turned into something easier to read.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Should change nothing on the user end, mostly to clean a few things up.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
test logs: 

![image](https://github.com/user-attachments/assets/45fa0933-e543-4135-a49b-686f913f0400)

proof gem changes the stored_gem value:

![image](https://github.com/user-attachments/assets/fa743858-1e23-4220-abf1-65d2d737c9f5)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->