(Hopefully) Fixed bad EOL in licenses.dm (#3013)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
A recent update to licenses.dm seems to have caused a problem by ending
the final line with a line feed, rather than a carriage return/line
feed, causing problems with tools (eg. VS Code is automatically
modifying the file for me, making it difficult to switch branches
properly). This PR should fix the issue by replacing the incorrect LF
with a CRLF as VS Code was automatically attempting to do.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Prevents development headaches. Doesn't modify any actual code, so there
should be no bugs associated with this change.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/9990dd56-de5f-4bdb-a747-c0ca938657e0)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->