Standardizes reagent container logic (adds features too plz read gyran!!) (#2976)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
works reagent containers to have common use interactions, sets fill
define, fixes pot insertion so there's no more special cases. or hopes
to...

I might still need to do more work on this but it's here as a proof of
concept until I think it's done.

edit:

- Standardizes all reagent container intent order to be consistent.
- adds `INTENT_FILL` define
- Removes the "this person cannot be milked" message, unneeded when
someone is standing there clothed.
- removes barrel update_icon (depreciated)
- prevents torches and their children types from being put into pots.
- allows pouring of reagent container's contents onto food (right click)
(video below)
- allows precise pouring of 1 action via right click


Creates two main procs for reagent containers to prevent repeating code:
`can_reagent_interact` and `do_reagent_interact`

`can_reagent_interact` is all the checks to see if it's possible to
transfer while `do_reagent_interact` is the functional aspect of the
transfers.


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
some interactions around reagent containers can be iffy and
inconsistent. This will try to unify it in as many was as possible.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/1cfd65bd-6ea2-424f-b407-02688d7c2012



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->