Alternative hand intent fix on forceMove (#3010)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
An alternative approach to the issue
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2999 tries to fix.

This one puts a simple check in the forceMove code itself to double
check if the location of the object is a human and if so to update the
intents.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Would be less tedious and you wouldn't have to think about keeping two
procs together should you use forceMove in these situations.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/870347ec-a35c-4cb4-955c-f5e9b5e1cc19


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->