You must wield polearms for their defense value (#3007)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
changes the calculation and adds a var to items called wieldedwdefense
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This should prevent issues with people using polearms in one hand as a
shield while fighting with another weapon. Use an actual shield.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->