Unanchors anvils, combat mode allows damage to them (#3002)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR lets anvils be moved, and damaged when combat mode is turned on,
so now they're no longer unmovable objects that can be used for
nefarious things.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Less exploits, and i guess a bit of QoL for smiths since now they can
customize their anvil placement a bit.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Cutting the anvil (It shows slightly damaged too, forgot to screenshot
it)

![image](https://github.com/user-attachments/assets/89015362-bdef-460e-91b5-6406ebab6074)
You can still place and repair weapons when not in combat mode.

![image](https://github.com/user-attachments/assets/ea0c5d69-b9d1-4f9b-83a6-07f48a3a4940)
Moving 

![image](https://github.com/user-attachments/assets/fc3c36d7-268b-4512-9117-f8e21993946f)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->