Adds zizo bane mushroom to code (#2986)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Currently not implemented in world (will have to be added to the bog
spawning code most likely)

Adds a mushroom named Zizo's bane. Stepping on it can cause spores to
launch out causing you to be sleepy. You can pluck the mushroom. Eating
it gives the same sleeping effect

- particle effect
- sound on step
- sprite for mushroom
- new debuff effect that causes sleep on remove

known issue:
Only issue I can see is standing in the cloud for too long would extend
the debuff (seems to be how stacking effects works) meaning you could
delay the sleep slightly by standing in the spore cloud or eating a
second mushroom however it would be something like an additional 5
seconds or so which perhaps isn't worth cheesing to rescue yourself.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
There needs to be more things that makes the rot feel scary. As a food
item this would also allow for potential RP elements that exploit making
people sleep. With being added to the bog flower spawn this would also
help it be a less common item that could cause people to explore the bog
for it in addition to paying a high price should they wish to sell one
to other players.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/1e84ebc3-3ab2-4751-9683-73da446568ac


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->