Hoardmaster changes (#3009)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Gives two health potions worth of reagents to the kidnapped victim when
they are kidnapped and put into the hoardmaster.
Admins see that this is causing problems with people dying from being
skull-cracked and left for dead in a spawn area.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Admin requested changes.
Why it's being requested by the admins: Having people with skull cracks
in the dock causes problems, and makes them into a deadite that spawn
camps people.

## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/dfd69b78-46d0-4d77-aa42-7b3da9fe1aa3)
Nothing breaks, and 2 bottles worth of health potion are given.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->