AP Greatweapon Strap (#2977)
## About The Pull Request
- - -
Additions:
- Introduces the 'greatweapon strap', pulled directly from AP. Almost a
1:1 copy paste. I think the icon path is the singular change, beyond the
roles that get it and some wording. Because it works. Original PR by
[AtoriBirb](https://github.com/AtoriBirb):
[Here](https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/814)
- - -
Roles Modified:
 - Knight Sentinel and Champion provided the GWS.
 - Retinue Halberdiers provided the GWS.
 - Grenzelhoft Mercenaries(Excluding Jaeger) provided the GWS.
 - Bandit Knight and Rogue Mage provided the GWS.
- - -
Crafting:
- You can craft the GWS, classified as a skill 3 sewing
recipe(Apprentice). It requires 2 leather + 1 rope.
- - -

What does this do? It allows weapons of a bulky size, such as halberds,
estocs and spears to be strapped directly to your back. To draw and stow
the weapons, it's a five second timer. This means it's almost entirely
inadvisable to be doing this in combat. Stats, for what little there are
with such an item, are also derived entirely from AP and unchanged. (The
time required, more-or-less.)

It kinda struck me real hard when I'd proposed this in OOC the other
day, only for someone to go "AP did that already goober". So I'm just
bringing over their work, as opposed to potentially making something
sloppy myself.
## Why It's Good For The Game

There's a significant, serious flaw to playing any role that frequently
makes use of two handed polearms / blades. Intentionally so, mind, but
no less obnoxious to the player. For such powerful weapons, you suffer
the requirement of having a hand slot always filled. Or, worse, you toss
it away when you move to properly handle someone on the ground. This PR
remedies that in a QoL manner that isn't just allowing someone to
instantly draw or stow these weapons. Such as with a back slot item
flag, for example.

It also allows characters with a lack of strength to properly carry
these same weapons about, as you can stow it while holding such in two
hands. Such as vermin/kobold bandits, who might loot such an item, or
typically carry a spear on their person and suddenly lose their strength
to hunger and just up and drop it.

That's not even touching on RP potential with such an item. No longer
are knights and halberdiers weirdos that don't put down their equipment
for a singular moment of their miserable lives.
## Proof of Testing (Required)
I really can't be bothered to get a video, so you get images instead.
Sorry.


![image](https://github.com/user-attachments/assets/9ec298d6-1efb-4fa4-9cc1-41a1da710c21)

![image](https://github.com/user-attachments/assets/d967e1c7-d389-4839-8241-cb109ed936e0)

![image](https://github.com/user-attachments/assets/e53b2761-fe6f-47cd-a07b-11e93e097ca4)


![image](https://github.com/user-attachments/assets/58e5a529-2001-45d3-a0c0-44f370f381fc)

![image](https://github.com/user-attachments/assets/46231369-4381-4315-b3b8-b80209bb7e21)

![image](https://github.com/user-attachments/assets/8430ba24-a27f-4303-b883-a671bec299ec)

---------

Co-authored-by: Useroth <37159550+Useroth@users.noreply.github.com>