Checks if the container is open for splashing (#3029)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Checks if the container is open for splashing
- adds 'The' for grammatical benefits.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Some things are closed containers so it helps keep logic consistent.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
if statement.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->