Gives Werewolves The Shaft. (#3030)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

This PR adds functioning shafts to werewolves. Now you can Were-ERP in
style!

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It was always a little sad not seeing the shaft when a werewolf was
using it - everybody else gets to have one, so why not werewolves?


## Proof of Testing (Required)

<details>
<summary> NSFW images </summary>

![Screenshot 2025-06-20
152558](https://github.com/user-attachments/assets/ffc9c8be-57b9-4169-a62a-37fc80fd3b0c)


![Screenshot 2025-06-20
152607](https://github.com/user-attachments/assets/2e0282b9-4f14-4908-8135-2f907bdf71ff)


![Screenshot 2025-06-20
152618](https://github.com/user-attachments/assets/1c2d6b6d-fbb3-4747-9df3-3422603e431e)


![Screenshot 2025-06-20
152624](https://github.com/user-attachments/assets/f176bd55-4b7f-478e-b679-c8d76c37aa6b)

</details>