Dye bin changes, adds color wheel, color sanitizing (#3025)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Does the following:

- Ports the AP color wheel choice for color selection as an option
- includes the sanitizing color proc to reduce colors too vibrant or too
dark in the choices
- removes berry requirement for dying
- removes do_after for dying

Only issue is that opening color wheels seems to make tabs flicker. It's
only for a moment and it happens on AP too so I imagine no one cares and
it's probably a byond quirk.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
makes dye bins a little less tedious to use if you have a lot of things
to color. also allows for expanding in future should we want to be able
to dye additional aspects of clothes.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/e39c0af2-31a0-44b3-9a32-e7f004439e4e


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->