(BOUNTY) Adds a Lion tail (#3033)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

![image](https://github.com/user-attachments/assets/8839f885-c6d5-4818-b0c5-f05cfe8a0c11)
Adds a tail to the game 
:)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More things
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/4a4f6d20-bbfc-4910-8054-899912189407)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->