Makes coif grayscale, gives default color (#3024)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes coif more compatible with dying colors into clothing. It's a
common item so it could do with it.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
more color
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
old and new:

![image](https://github.com/user-attachments/assets/0078f450-21b0-4308-82cb-2a3aefad5799)

color tests:

![image](https://github.com/user-attachments/assets/428e4ece-957f-4ae6-9de3-fee4453febcb)

![image](https://github.com/user-attachments/assets/3037a7a7-be08-4737-af1e-dbe2862a4384)

![image](https://github.com/user-attachments/assets/61538515-87e9-49c1-aa5b-a375e9758412)

More color examples in the Azure PR:
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/2823
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->