Explains Towner houses to Towners (#3015)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds an extra line to the Towner greet message, saying that you can
claim a house as your own by clicking on it.

## Why It's Good For The Game

Addresses Issue #1536 .
This should really be written out somewhere, because it isn't in any
documentation, or anything to my knowledge. This is a simple little fix
until a future rework of Towners/families/houses etc.

## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/deeb4a30-4479-480b-b55a-0aa234a74caf)

---------

Co-authored-by: zoey <91065935+proxiwastaken@users.noreply.github.com>