Start of loadout framework (#3020)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds the loadout items framework from Azurepeak to Ratwood. Adds many
clothing items and stuff like cards or dice to the loadouts. Easily
modifiable so if we decide to add more clothing to it or remove items
from it

## Why It's Good For The Game

Players have been asking for this for ages and i was finally approved to
work on these changes. Adds customization to character in with extra
clothing they may want for their role


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/c02a223b-577c-4f0c-9189-6895fef42893)

![image](https://github.com/user-attachments/assets/0582705f-e60c-4e00-a275-a3da529274bc)