Added Aspects system & initial aspects (#3046)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds the basic infrastructure for Aspects, alongside a few
initial aspects. These are:
**Magocracy**: all nobles spawn with some arcane skill,
Prestidigitation, and two randomly-selected spells (excluding the
13-cost spells). Any character who spawns with the Arcane skill becomes
a noble.
**Magic Town**: all player characters gain the benefits of _Magocracy_.
**Woman's World**: Women gain a +1 to strength and constitution, men
gain a -1 to strength and constitution. Changes are inverted for dark
elves.
**The Rot Abates**: All player characters are immune to the rot.
**Gifted**: All player characters' highest non-legendary skill becomes
Legendary instead. If there's a tie, one of the highest non-legendary
skills is chosen randomly.

In the lobby, once the map has loaded, there is a 35% chance that the
game will initiate a vote for which Aspect to play with. Two aspects
will be chosen from a weighted list as voting options; a third option,
"None", is present for cases in which players would rather play without
any of the presented options.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Aspects were originally a LFWB invention, as I understand; and while I
don't like the sound of every aspect I've been told about, particularly
the more intensely disruptive ones, some of them nevertheless spurred my
imagination (_Woman's World_, notably, is directly inspired by one I was
told about). Personally, I like to imagine aspects as small permutations
on the setting - a version of the setting where everyone is gifted with
magic, or where magic is considered a sign of nobility; where women are
stronger than men, where people are better at the things they do.

Per this interpretation, I think - alongside adding gameplay variety -
they equally offer something for those so inclined to roleplay around.
Meanwhile, the 65% chance of Aspects not being applied at all keeps them
from becoming _expected_; and the vote, rather than allowing the Duke to
choose as in LFWB's original implementation, allows players to
effectively veto those aspects they particularly dislike, without any
need to immediately remove them.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
No screenshots here, but I've tested spawning in - both on round start
and as a late joiner - for each implemented aspect, and all of them
behave as expected. This should be sufficient for now, since all aspect
effects currently are applied to characters when they spawn.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->